fix(protocol-designer): unify navigation bar of pd
unify navigation bar of pd

close RQA-3761
