+++ protocol-designer/src/molecules/index.ts
@@ -2,6 +2,5 @@ export * from './CheckboxExpandStepFormField'
 export * from './CheckboxStepFormField'
 export * from './DropdownStepFormField'
 export * from './InputStepFormField'
-export * from './SettingsIcon'
 export * from './ToggleExpandStepFormField'
 export * from './ToggleStepFormField'
