+++ protocol-designer/src/organisms/NavigationBar/__tests__/NavigationBar.test.tsx
@@ -2,17 +2,17 @@ import { describe, it, vi, beforeEach, expect } from 'vitest'
 import { fireEvent, screen } from '@testing-library/react'
 import { MemoryRouter } from 'react-router-dom'
 
-import { i18n } from '../assets/localization'
-import { renderWithProviders } from '../__testing-utils__'
-import { NavigationBar } from '../NavigationBar'
-import { getHasUnsavedChanges } from '../load-file/selectors'
-import { toggleNewProtocolModal } from '../navigation/actions'
-import { SettingsIcon } from '../molecules'
+import { i18n } from '../../../assets/localization'
+import { renderWithProviders } from '../../../__testing-utils__'
+import { getHasUnsavedChanges } from '../../../load-file/selectors'
+import { toggleNewProtocolModal } from '../../../navigation/actions'
+import { SettingsIcon } from '../../SettingsIcon'
+import { NavigationBar } from '..'
 
-vi.mock('../molecules')
-vi.mock('../navigation/actions')
-vi.mock('../file-data/selectors')
-vi.mock('../load-file/selectors')
+vi.mock('../../SettingsIcon')
+vi.mock('../../../navigation/actions')
+vi.mock('../../../file-data/selectors')
+vi.mock('../../../load-file/selectors')
 const render = () => {
   return renderWithProviders(
     <MemoryRouter>
