+++ protocol-designer/src/organisms/SettingsIcon/__tests__/SettingsIcon.test.tsx
