+++ protocol-designer/src/organisms/index.ts
@@ -17,10 +17,12 @@ export * from './Kitchen'
 export * from './KnowledgeLink'
 export * from './LabwareOnDeck'
 export * from './LabwareUploadModal'
+export * from './NavigationBar'
 export * from './PipetteInfoItem'
 export * from './Portal'
 export * from './ProtocolNavBar'
 export * from './SelectWellsModal'
+export * from './SettingsIcon'
 export * from './SlotDetailsContainer'
 export * from './SlotInformation'
 export * from './TipPositionModal'
